Moved feature-info code outside settings tab
